refactor audio player
