+++ src/components/audio/index.ts
@@ -1,0 +1,4 @@
+export { AudioPlayer } from "./AudioPlayer";
+export { SmartSpeedPlayer } from "./SmartSpeedPlayer";
+export { useAudioPlayer } from "./useAudioPlayer";
+export { useTts } from "./useTts";
